Refactoring state_space functions.
