+++ tests/test_function_wrapping.py
@@ -1,9 +1,6 @@
-import functools
-import inspect
-from functools import partial
-
 import jax.numpy as jnp
 import numpy as np
+from dcegm.process_functions import determine_function_arguments_and_partial_options
 from dcegm.process_model import _convert_params_to_dict
 from jax import vmap
 from toy_models.consumption_retirement_model.utility_functions import (
@@ -45,7 +42,7 @@ def test_function_wrapping(load_example_model):
         [0.0, 0.69314718, 1.09861229, 1.38629436, 1.60943791, 1.79175947], dtype=float
     )
 
-    util_processed = simple_wrapping(
+    util_processed = determine_function_arguments_and_partial_options(
         utiility_func_log_crra, options, {"ltc": np.array([0])}
     )
 
@@ -57,51 +54,3 @@ def test_function_wrapping(load_example_model):
 
 def util_wrap(state_dict, params, util_func):
     return util_func(**state_dict, params=params)
-
-
-def simple_wrapping(func, options, exog_state_space):
-    signature = set(inspect.signature(func).parameters)
-
-    exogenous_processes_names = set(
-        options["state_space"]["exogenous_processes"].keys()
-    )
-
-    exogs_in_signature = list(signature.intersection(exogenous_processes_names))
-    signature_kwargs_without_exog = list(signature.difference(exogs_in_signature))
-
-    exog_mapping = create_exog_mapt(exogs_in_signature, exog_state_space)
-
-    options_processed_func = partial_options(func, signature, options)
-
-    @functools.wraps(func)
-    def processed_func(**kwargs):
-        exog_kwargs = exog_mapping(kwargs["global_exog"])
-
-        other_kwargs = {key: kwargs[key] for key in signature_kwargs_without_exog}
-
-        return options_processed_func(**exog_kwargs, **other_kwargs)
-
-    return processed_func
-
-
-def partial_options(func, signature, options):
-    if "options" in signature:
-        return partial(func, options=options)
-    else:
-        return func
-
-
-def create_exog_mapt(exogs_in_signature, exog_state_space):
-    if len(exogs_in_signature) > 0:
-
-        def exog_mapping(x):
-            return {
-                exog: jnp.take(exog_state_space[exog], x) for exog in exogs_in_signature
-            }
-
-    else:
-
-        def exog_mapping(x):
-            return {}
-
-    return exog_mapping
